requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
baron==0.10.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
debugpy==1.6.6
exceptiongroup==1.1.0
getgauge==0.3.17
grpcio==1.51.3
h11==0.14.0
idna==3.4
outcome==1.2.0
protobuf==3.19.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
PySocks==1.7.1
redbaron==0.9.2
requests==2.28.2
rply==0.7.8
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
urllib3-secure-extra==0.1.0
wsproto==1.2.0
